Add pymarkdownlnt to check_examples.
