+++ tools/check_examples/Pipfile
@@ -15,3 +15,4 @@ black = "*"
 pylint = "*"
 pydocstyle = "*"
 mypy = "*"
+pymarkdownlnt = "*"
