+++ tools/check_examples/check_basic.py
@@ -1,7 +1,7 @@
 """Check for basic requirements of example repositories."""
 
 from pathlib import Path
-from common import file_exists
+from common import file_exists, list_files_of_suffix, execute
 from terminal_utils import iprint_fail, iprint_pass
 
 
@@ -18,4 +18,18 @@ def check_basic(directory: Path, indent: int) -> int:
         else:
             iprint_pass(f"{file} found.", indent)
 
+    return return_code
+
+
+def check_markdown(directory: Path, indent: int) -> int:
+    """Check markdown files."""
+    return_code = 0
+    markdown_files = list_files_of_suffix(directory, ".md")
+    for file in markdown_files:
+        filename = str(file.relative_to(directory))
+        return_code |= execute(
+            ["pipenv", "run", "pymarkdownlnt", "--config", "pymarkdownlnt.toml", "scan", filename],
+            directory,
+            indent,
+        )
     return return_code
