+++ tools/check_examples/pymarkdownlnt.toml
@@ -1,0 +1,2 @@
+[plugins.md013]
+line_length = 100
